utilisation de la place du semaphore et changement de la fonction consommateur

+++ libfractal/fractal.h
@@ -7,7 +7,7 @@ struct fractal {
   int height;
   double a;
   double b;
-  double *value;// a voir si c'est pas plutot des int.
+  int *value;// a voir si c'est pas plutot des int.
 };
 
 /*
